Tidied up as per comments from Pete
